requirements.txt
google-cloud
google-cloud-pubsub
errbot